feat: implement finding modules when it is a dependency
This implements finding the flux module when it is not the main module.
When a replace is used with a file path, `pkg-config` will compile it
the same as if it were the main module.

This also implements downloading the module to the module cache to
perform the build there. This build will fail though because the
directory is read only.