fix: modify path correctly when pkg-config is installed in multiple locations
If `pkg-config` was installed in multiple places on the path, it would
be possible for one of those `pkg-config` executables to invoke the
other, then for that one to invoke the original, in an infinite loop.

So the path resolution has now been modified to find the current
executable on the path and, if present, it will remove all entries
before it. This way, it avoids triggering an infinite loop.