fix: build module when it is in a vendor directory or the module cache (#6)
The `pkg-config` executable made the erroneous assumption it would be
called within the directory of the project that was invoking it rather
than inside of the package that was using it.

The `pkg-config` executable can now recognize when it is called from
within the module cache and it can find the version by using the
directory filepath instead of relying on git for those situations.

It also is now able to recognize when it is within a vendor directory.
After it finds the module root, it looks to see if there is another
module root in the path with a vendor directory. If there is, it will
use the original project instead of the vendored directory and it will
proceed with the `go mod download` workflow.