fix: hash the libraries to produce a unique file per source input (#21)
This changes `pkg-config` to produce a unique hashed linked output and
configures it to output a link command that includes that hash. This is
done because Go will hash the output of `pkg-config` and it will hash
multiple builds.

For example, it is possible to use `pkg-config` on one build and get a
valid output. Then you make a modification and use it again and get
another valid output. If you revert back to the first build input,
go will believe the library has already been built and won't run
`pkg-config` again. It will link against the output from the second
build.

This is because go's invocation of `pkg-config` always assumes that
nothing changes. By adding these hashes, go will use the first
`pkg-config` invocation's output and link against the correct library.

This mostly affects local builds because local builds are the only
circumstance where `pkg-config` will overwrite the old file. Module and
vendor builds aren't affected because they produce a unique directory
output.