feat: add support for modversion flag
The `modversion` flag will report available versions for a comma
separated list of package names.

Today, the wrapper doesn't recognize this flag. This means while
building `influxdb_iox`, its dependency on `openssl-sys` will fail when
it tries to run `pkg-config --modversion openssl`.

This diff adds support for this flag so that `influxdb_iox` can be
compiled without having to sidestep this wrapper.